OK, someone *seriously* fucked up the <gtkmm/papersize.h> header introduced with gtkmm 2.10.  You won't believe your eyes.  Static objects with runtime initializers.  In a header file.  In a public library header file.  Which is pulled by default from <gtkmm.h>.  That's a BIG Ooops.
Since I hate the idea of playing along with this crass API fuckup, I replaced the inclusion of <gtkmm.h> with a subset of individual gtkmm include files.  This is a bit of a pain since the list of header dependencies can grow quite long, but it's only for two files and thus manageable.