meshloader: Do not throw in thread completion handler
There is no point in re-throwing the exception collected from the
finished thread directly in the Dispatcher signal handler. User code
has no means to catch that other than by installing a global glibmm
exception handler. It is sufficient to re-throw on the first attempt
to access the loader result, which the code already does anyway.