Compile mesh data to binary at build time
Instead of parsing the Collada XML resource at run-time, use
a build-time utility to compile the vertex and index data into
a form that can be fed directly to OpenGL. This significantly
simplifies and speeds up the mesh loading on application start.
It also turns out that the uncompressed binary data is actually
slightly smaller than the compressed Collada file.

Note that this downgrades assimp from a run-time to a build-time
dependency.

Further modifications are required to get the new build tool
working when cross-compiling.