Pre-compress RGTC wood texture
Pre-generating mipmaps and pre-compressing the wood texture makes the
texture loading code path so much simpler. The downside is that the
compression codec RGTC1 has to be hard-coded, so this currently won't
work on most OpenGL ES implementations. However, this could be dealt
with later by adding a EAC compressed variant.