Use row-major 3x4 matrices for model-view transform
Switching to row-major layout for the model-view transformation in
the vertex shader allows one matrix column to be dropped entirely.
Additionally, widen the position and normal vertex attributes to
full 4-vectors, taking advantage of implicit extension by w=1 for
the position vectors. Taken together, these changes further reduce
both the instruction count and the cycle count of the vertex shader
on Intel hardware, and likely on others platforms too.