Transform grid lines to triangles in screen space
Classic OpenGL line drawing is inflexible, varies in quality between
implementations, and isn't particularly fast either. To overcome these
problems, employ a geometry shader to transform each line segment to
a pair of triangles forming a rectangle in screen space. The resulting
long and thin triangles aren't a good fit for the rasterizer, but the
performance is no worse than drawing plain old lines.

Use the gained flexibility to draw wider lines, and to scale the line
width on HiDPI displays along with the UI elements.