Explicitly bind fragment shader outputs
On desktop OpenGL, bind the output color of all fragment shaders to
color number 0 before linking. The spec is not entirely clear if this
is actually necessary if there is only a single output, but better be
safe than sorry.

The OpenGL ES spec on the other hand does specify that single outputs
are always bound to color number 0 implicitly. Which is just as well,
given that the ES API does not support binding the color number
programatically!