Fix calculation of rotation between vectors
Quaternions are constructed from the sine and cosine of the half
angle, whereas the vector cross and dot products return a function
of the full angle. The constructed quaternion needs to be adjusted
accordingly.

Introduce a dedicated method for calculating the rotation between
two vectors, and do the adjustment there.