Clarify matrix multiply aliasing restrictions
Also rearrange the SIMD fallback implementation of matrix multiply
in such a way that the internal copy is only needed if the result
aliases the right input matrix, rather than the left one. Essentially
this means that the copy path will never be executed except in the
degenerate case of squaring a matrix in place.