Move some vector operations to separate functions
Favor free-standing functions over methods for basic math operations
like magnitude, invert, etc. Also demote the dot and cross products
from operators to ordinary functions. For methods that have both an
in-place version and a non-modifying one, keep the former and turn
the latter into a non-member function of the same name.

These changes further align the syntax with GLSL.