Get rid of custom memory allocation for SSE
The custom implementations of operator new and delete added a lot of
complexity, and suffered from code rot since they are unused on Linux
x86-64 anyway. Now that C++17 introduced automatic aligned allocation
for types with extended alignment requirements, we can just rely on
that for the platforms that might need it.

Also drop related configuration checks for system-level aligned memory
allocation support.