Get rid of the bit cube z coordinate inversion
Due to an unfortunate historical accident, the bit cube structure
used a left-handed coordinate system with the z axis pointing away
from the viewer. This was even inconsistently applied; e.g. shifts
worked in the left-handed system whereas rotations corrected the z
direction to simulate a right-handed coordinate system. At various
points throughout the code, the z coordinate had to be inverted to
accommodate the difference to OpenGL's right-handed conversion,
making the logic hard to follow.

Resolve the situation by adopting the right-handed convention for
the bit cube. This necessitated changes to both the Soma cube data
and the mesh data in order to realign untransformed cube pieces to
the lowest z coordinate in the 3x3x3 cell grid. The animation logic
and transforms had to be adjusted as well. However, the logic and
transformations are much easier to follow now.