Construct static bit cubes from list of cell indices
Instead of a list of booleans for every cell in a bit cube, change
the compile-time constant bit cube constructor to take a list of cell
index triplets describing all occupied cells in the bit cube. This is
much more compact and removes the hard dependency on the bit order.